local_file_server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0
